Use "experimental" class fields syntax
In order to bind "this" to the functions automatically and hence, reduce
redundant code like:

    addMove={move => this.addMove(move)}

to:

    addMove={this.addMove}

This became possible because we have changed "addMove" method's
declaration to use class fields syntax, instead of class function
syntax. Class fields syntax (which is a "function expression")
automatically binds "this" to the function, whereas class function
syntax (which is a "function declaration") DOES NOT bind "this" to the
function. If you want to bind "this" to a class function syntax, you
either have to call "bind" explicitly on it, or use the redundant method
demonstrated in the first code block above.